Almost

+++ prom_movil.c
@@ -8,7 +8,7 @@ double* obtener_promedio_movil(int* arreglo, size_t n, size_t k){
     double* result = calloc(n, sizeof(double));
     for (int i = 0; i < n; ++i) {
         int cont = 0;
-        for (int j = (i - (int)k); j <= (i+k); ++j) {
+        for (int j = (i - (int)k); j <= (i+(int)k); ++j) {
             //Entra recien en i=1 (Ignora el i=0)
             if(j < 0 || j >= n)
                 continue;
